requirements.txt
streamlit==1.41.1
google-generativeai==0.8.4
pydantic==2.10.5
python-dotenv==1.0.1
langdetect==1.0.9
pytest==8.3.4
